+++ src/basic/hooks/useProducts.ts
@@ -1,0 +1,126 @@
+import { useState, useCallback } from "react";
+import { ProductWithUI } from "../models";
+import { useLocalStorage } from "./useLocalStorage";
+
+// 초기 데이터
+const initialProducts: ProductWithUI[] = [
+  {
+    id: "p1",
+    name: "상품1",
+    price: 10000,
+    stock: 20,
+    discounts: [
+      { quantity: 10, rate: 0.1 },
+      { quantity: 20, rate: 0.2 },
+    ],
+    description: "최고급 품질의 프리미엄 상품입니다.",
+  },
+  {
+    id: "p2",
+    name: "상품2",
+    price: 20000,
+    stock: 20,
+    discounts: [{ quantity: 10, rate: 0.15 }],
+    description: "다양한 기능을 갖춘 실용적인 상품입니다.",
+    isRecommended: true,
+  },
+  {
+    id: "p3",
+    name: "상품3",
+    price: 30000,
+    stock: 20,
+    discounts: [
+      { quantity: 10, rate: 0.2 },
+      { quantity: 30, rate: 0.25 },
+    ],
+    description: "대용량과 고성능을 자랑하는 상품입니다.",
+  },
+];
+
+export const useProducts = () => {
+  const [products, setProducts] = useLocalStorage("products", initialProducts);
+
+  // 상품 폼 관련 상태
+  const [editingProduct, setEditingProduct] = useState<string | null>(null);
+  const [showProductForm, setShowProductForm] = useState(false);
+  const [productForm, setProductForm] = useState({
+    name: "",
+    price: 0,
+    stock: 0,
+    description: "",
+    discounts: [] as Array<{ quantity: number; rate: number }>,
+  });
+
+  const addProduct = useCallback(
+    (newProduct: Omit<ProductWithUI, "id">) => {
+      const product: ProductWithUI = {
+        ...newProduct,
+        id: `p${Date.now()}`,
+      };
+      setProducts((prev) => [...prev, product]);
+    },
+    [setProducts]
+  );
+
+  const updateProduct = useCallback(
+    (productId: string, updates: Partial<ProductWithUI>) => {
+      setProducts((prev) =>
+        prev.map((product) =>
+          product.id === productId ? { ...product, ...updates } : product
+        )
+      );
+    },
+    [setProducts]
+  );
+
+  const deleteProduct = useCallback(
+    (productId: string) => {
+      setProducts((prev) => prev.filter((p) => p.id !== productId));
+    },
+    [setProducts]
+  );
+
+  const startEditProduct = useCallback((product: ProductWithUI) => {
+    setEditingProduct(product.id);
+    setProductForm({
+      name: product.name,
+      price: product.price,
+      stock: product.stock,
+      description: product.description || "",
+      discounts: product.discounts || [],
+    });
+    setShowProductForm(true);
+  }, []);
+
+  const resetProductForm = useCallback(() => {
+    setProductForm({
+      name: "",
+      price: 0,
+      stock: 0,
+      description: "",
+      discounts: [],
+    });
+    setEditingProduct(null);
+    setShowProductForm(false);
+  }, []);
+
+  return {
+    // 상태
+    products,
+    editingProduct,
+    showProductForm,
+    productForm,
+
+    // 액션
+    addProduct,
+    updateProduct,
+    deleteProduct,
+    startEditProduct,
+    resetProductForm,
+
+    // 세터
+    setEditingProduct,
+    setShowProductForm,
+    setProductForm,
+  };
+};
